Refactor customer methods in MainService
Reorganized the methods in the MainService, separating the process of adding new customers and updating existing customer events. The 'addCustomer' method now generates a new ID, sets it to the customer, and then uses 'updateCustomerEvent' to send the data to the Kafka topic. This improves readability and separation of concerns.