Remove FRIEND_IP check from entrypoint.sh
The condition to check the existence of FRIEND_IP environment variable was removed from the file entrypoint.sh. This IP address check was found to be unnecessary during the validation process and was causing an error when missing, which led to functional issues in the operation of the messaging server.