Add Docker build steps for messaging rest images
New steps have been added to the Docker build workflow to create images for the messaging rest sender and receiver. Besides, the context and file for the event sourcing producer image have been updated to use correct paths. This ensures that the correct Docker images are built and used for the messaging components in the microservices system.