Update service IP URL in application.properties
The service IP with address 192.168.1.3 has been modified in the application.properties file, under the circuit_breaker/proxy directory. The URL has been updated to include the actual path of the real service. Previously, it was pointing to the network node instead of the actual service path, which is now corrected.