Correct proxy jar file path in entrypoint.sh
The path pointed to find proxy-0.0.1.jar file in the entrypoint.sh script from the "circuit_breaker" directory is updated. The jar file is referenced from "/home/proxy_user/" rather than "/home/proxy". This change resolves the issue of the script not being able to locate the file in the previous path.