Update GitHub Actions to upload and push Docker images
This commit updates the GitHub Actions workflow with steps to upload the artifacts of each microservice and then push their Docker images to Docker Hub. This comprehensive update aims to automate the process of building, uploading, and deploying all the Docker images for the microservices involved in the lab environment.