Remove docker build image GitHub Action
The GitHub Action workflow that is responsible for building docker images daily, has been deleted. This change won't affect the core logic of the project but it does affect how we generate and update our Docker images. Discontinuing this approach means a new workflow for image building should be implemented if necessary.